+++ addOns/reports/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- The Automation Framework progress to the report data when run via an AF job.
+
 ### Fixed
 - Correct error messages of the Automation Framework job.